Re-organise controllers to increase consistency.
