+++ TourOperator/Views/Auth/Login.cshtml
@@ -15,5 +15,5 @@
         <input class="btn btn-primary" type="submit" value="Login"/>
     </form>
     
-    <a class="btn btn-secondary" asp-area="" asp-controller="View" asp-action="Register">Create An Account</a>
+    <a class="btn btn-secondary" asp-area="" asp-controller="Auth" asp-action="Register">Create An Account</a>
 </div>
