+++ TourOperator/Views/Auth/Register.cshtml
@@ -7,7 +7,7 @@
 <div class="text-center">
     <h1 class="display-4">Register</h1>
     
-    @using(Html.BeginForm("Register", "View", FormMethod.Post)) {
+    @using(Html.BeginForm("Register", "Auth", FormMethod.Post)) {
         <div>
             <label asp-for="Username"></label><br/>
             <input asp-for="Username"/>
