+++ TourOperator/Views/Booking/AllBookings.cshtml
@@ -42,7 +42,7 @@
                 </td>
                 
                 <td>
-                    <a class="btn btn-secondary" asp-controller="View" asp-action="BookingInfo" asp-route-bookingId="@booking.Id">View</a>
+                    <a class="btn btn-secondary" asp-controller="Booking" asp-action="BookingInfo" asp-route-bookingId="@booking.Id">View</a>
                 </td>
             </tr>
         }
