+++ TourOperator/Views/Shared/_Layout.cshtml
@@ -49,7 +49,7 @@
                 }
                 else
                 {
-                    <a class="nav-link text-dark" asp-area="" asp-controller="View" asp-action="Login">Login</a>
+                    <a class="nav-link text-dark" asp-area="" asp-controller="Auth" asp-action="Login">Login</a>
                 }
             </div>
         </div>
