+++ TourOperator/Views/View/Customer.cshtml
@@ -7,7 +7,7 @@
 <div class="text-center">
     <h1 class="display-4">Customer</h1>
     
-    <a asp-controller="View" asp-action="Bookings" class="btn btn-secondary">Bookings</a>
+    <a asp-controller="Booking" asp-action="AllBookings" class="btn btn-secondary">Bookings</a>
     
     <form method="post" asp-controller="View" asp-action="Customer">
         <div>
